+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+## Added
+- A server configuration to blacklist recipes that shouldn't be inverted.
+The default blacklist has some recipes to avoid item duplication.
+If you find more recipe that lead to duplication, please tell me, so I can add them to the default blacklist.
+
+## Changed
+- The recipe is now easier
+- The texture of the uncrafter has been changed to math the crafting recipe
+
 ## 1.17.1-0.1.0
 
 Initial beta release